requirements.txt
aniso8601==9.0.1
antiorm==1.2.1
cffi==1.15.0
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
db==0.1.1
dnspython==2.2.1
document==1.0
email-validator==1.1.3
Flask==2.1.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jwt==1.3.1
MarkupSafe==2.1.1
mongoengine==0.24.1
pycparser==2.21
PyJWT==2.3.0
pymongo==4.1.1
pytz==2022.1
six==1.16.0
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0

python-dotenv~=0.20.0